修改更新脚本

+++ scripts/7_upgrade.sh
@@ -65,7 +65,11 @@ function upgrade_service() {
 
     echo -ne "更新升级文件 ... "
     package_name=`ls -l ${PROJECT_DIR}/data/packages/ | grep -v total | awk '{print $9}'`
-    rm -rf ${KUBEOPS_DIR}/data/packages/${package_name}
+    if [ ! ${package_name} ]; then
+        echo -ne "无k8s安装包 跳过此步骤 ... "
+    else
+      rm -rf ${KUBEOPS_DIR}/data/packages/${package_name}
+    fi
     \cp -rf ${PROJECT_DIR}/* ${KUBEOPS_DIR}/
     chmod -R 777 ${KUBEOPS_DIR}/data
     echo "[OK]"
